separate modelstructure and training code, save the model after training, create a separate inference file, update gitignore
