+++ .gitignore
@@ -19,6 +19,9 @@ env/
 # VSCode or PyCharm
 .vscode/
 .idea/
+log_profiler/
 
 # bin
 *.bin
+*.pt
+*.pkl
