+++ models/ReadMe.txt
@@ -1,0 +1,2 @@
+This is a fine-tuned GPT-2 model on the FineWeb10B dataset.
+To load the model, use `model.load_state_dict(torch.load('gpt2_fineweb_124M.pt'))`.
