+++ models/config.json
@@ -1,0 +1,7 @@
+{
+    "block_size": 1024,
+    "n_embd": 768,
+    "n_head": 12,
+    "n_layer": 12,
+    "vocab_size": 50304
+}
